+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fix
 
 - escaping of `\{\{2\}\}` did not work (Anweber/vscode-httpyac#244)
+- support only `http-client.env.json` and `http-client.private.env.json` files for intellij environments (Anweber/vscode-httpyac#245)
 
 ## [6.10.0] (2023-11-13)