+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - support only `http-client.env.json` and `http-client.private.env.json` files for intellij environments (Anweber/vscode-httpyac#245)
 - add support to set requestUnauthorized to OAuth2 Requests (#605)
 - allow intercepting and modification of OAuth2 Request before sending using `oauht2_interceptRequest` (#605)
+- allow empty string in input and list provider (AnWeber/vscode-httpyac#247)
 
 ## [6.10.0] (2023-11-13)